Move validator services inside spark job directory
