+++ piezo_web_app/PiezoWebApp/run_piezo.py
@@ -9,9 +9,9 @@ from PiezoWebApp.src.handlers.get_logs import GetLogsHandler
 from PiezoWebApp.src.handlers.submit_job import SubmitJobHandler
 from PiezoWebApp.src.services.kubernetes.kubernetes_adapter import KubernetesAdapter
 from PiezoWebApp.src.services.spark_job.manifest_populator import ManifestPopulator
-from PiezoWebApp.src.services.validator.validation_rules import ValidationRules
+from PiezoWebApp.src.services.spark_job.validation import ValidationRules
 from PiezoWebApp.src.services.spark_job.spark_job_service import SparkJobService
-from PiezoWebApp.src.services.validator.argument_validation_service import ArgumentValidationService
+from PiezoWebApp.src.services.spark_job.validation import ArgumentValidationService
 from PiezoWebApp.src.utils.route_helper import format_route_specification
