+++ piezo_web_app/PiezoWebApp/src/services/spark_job/spark_job_property.py
@@ -1,4 +1,4 @@
-from PiezoWebApp.src.services.validator import argument_validator
+from PiezoWebApp.src.services.spark_job.validation import argument_validator
 
 
 class SparkJobProperty:
